train/requirements.txt
numpy==1.19.5
Pillow==8.4.0
pytorch==1.7.1
torchvision==0.8.2
tqdm==4.48.2
